test: move fake to test file and remove duplicated test
