+++ cmd/namespace/namespace.go
@@ -18,16 +18,9 @@ import (
 
 	contextCMD "github.com/okteto/okteto/cmd/context"
 	"github.com/okteto/okteto/cmd/utils"
-	"github.com/okteto/okteto/internal/test/client"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/okteto/okteto/pkg/types"
 	"github.com/spf13/cobra"
-	"k8s.io/client-go/kubernetes"
-	"github.com/okteto/okteto/internal/test"
-	"github.com/okteto/okteto/pkg/k8s/ingresses"
-	"k8s.io/client-go/rest"
-	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
-
 )
 
 // NamespaceCommand has all the namespaces subcommands
@@ -70,35 +63,4 @@ func Namespace(ctx context.Context) *cobra.Command {
 	cmd.AddCommand(Sleep(ctx))
 	cmd.AddCommand(Wake(ctx))
 	return cmd
-}
-
-type fakeK8sProvider struct {
-	k8sClient kubernetes.Interface
-}
-
-func (p *fakeK8sProvider) Provide(_ *clientcmdapi.Config) (kubernetes.Interface, *rest.Config, error) {
-	return p.k8sClient, nil, nil
-}
-
-func (*fakeK8sProvider) GetIngressClient() (*ingresses.Client, error) {
-	return nil, nil
-}
-
-func newFakeContextCommand(c *client.FakeOktetoClient, user *types.User) *contextCMD.ContextCommand {
-	return &contextCMD.ContextCommand{
-		OktetoClientProvider: client.NewFakeOktetoClientProvider(c),
-		K8sClientProvider:    test.NewFakeK8sProvider(nil),
-		LoginController:      test.NewFakeLoginController(user, nil),
-		OktetoContextWriter:  test.NewFakeOktetoContextWriter(),
-	}
-}
-
-func NewFakeNamespaceCommand(okClient *client.FakeOktetoClient, k8sClient kubernetes.Interface, user *types.User) *NamespaceCommand {
-	return &NamespaceCommand{
-		okClient: okClient,
-		ctxCmd:   newFakeContextCommand(okClient, user),
-		k8sClientProvider: &fakeK8sProvider{
-			k8sClient: k8sClient,
-		},
-	}
 }
